requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-background-tasks==1.2.8
idna==3.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
